Fix 0x00 transmit issue, details see comment
The last commit used the SBI instruction to toggle the TX output pin. If no toggling is required, this instruction was skipped. Executing SBI takes 2 cycles, but skipping only 1 cycle; after 9 consecutive zeros the error was large enough to cause errors on the receiver.

The new code uses the out instruction instead, it always executes in 1 cycle. The flash requirements are identical